Rename setupfile.txt to .json and stop tracking
Rename the settings tile to setupfile.json and remove it from the repo
(it never should have been tracked). Add it to .gitignore to not add by
accident. Because now the code doesn't run straight out of the box,
setup.py has been modified to build an example setupfile if none exists.